requirements.txt
python-telegram-bot
requests
yt-dlp
